move pretraining logger fn to training.utils
